[backend] fix import cycle
