+++ opencti-platform/opencti-graphql/src/schema/schemaUtils.js
@@ -53,6 +53,13 @@ export const shortHash = (element) => {
   return hash.slice(0, 8);
 };
 
+export const isValidDate = (stringDate) => {
+  const dateParsed = Date.parse(stringDate);
+  if (!dateParsed) return false;
+  const dateInstance = new Date(dateParsed);
+  return dateInstance.toISOString() === stringDate;
+};
+
 const pascalize = (str) => {
   return str
     .match(/[a-z0-9]+/gi)
